move mongodb provider into Bridge
