+++ src/Symfony/Bridge/MongoDB/Security/Acl/MutableAclProvider.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Symfony\Component\Security\Acl\MongoDB;
+namespace Symfony\Bridge\MongoDB\Security\Acl;
 
 use Doctrine\Common\PropertyChangedListener;
 use Doctrine\MongoDB\Database;
