+++ tests/Symfony/Tests/Bridge/MongoDB/Security/Acl/AclProviderBenchmarkTest.php
@@ -1,9 +1,9 @@
 <?php
 
-namespace Symfony\Tests\Component\Security\Acl\MongoDB;
+namespace Symfony\Tests\Bridge\MongoDB\Security\Acl;
 
 use Doctrine\MongoDB\Connection;
-use Symfony\Component\Security\Acl\MongoDB\AclProvider;
+use Symfony\Bridge\MongoDB\Security\Acl\AclProvider;
 use Symfony\Component\Security\Acl\Domain\PermissionGrantingStrategy;
 use Symfony\Component\Security\Acl\Domain\ObjectIdentity;
 use Symfony\Component\Security\Acl\Domain\RoleSecurityIdentity;
